Separate vue-i18n from vuejs - Create new module - clean up vuejs config
