+++ modules/flamingo-carotene-vue-i18n/.gitignore
@@ -1,0 +1,8 @@
+.DS_Store
+.cache
+.idea
+.gradle
+.temp
+node_modules
+npm-debug.log
+yarn-error.log
