+++ modules/flamingo-carotene-vue-i18n/lib/handler/build.js
